requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
jupyter>=1.0.0
matplotlib>=3.5.0
numpy>=1.20.0 